Separate core and ui for GetterSetterCorrectionSubProcessor
